Changed mouse positions from vec2 to ivec2 values. Fixed some bugs in the InputManager related to delta positions. Added a commented log line for orientation in the PlayerControlSystem.

+++ RootEngine/InputManager/Include/InputInterface.h
@@ -28,8 +28,8 @@ namespace RootEngine
 			virtual KeyState::KeyState GetKeyState(MouseButton::MouseButton p_button) = 0;
 
 			// Returns the global and delta mouse position
-			virtual glm::vec2 GetGlobalMousePos() = 0;
-			virtual glm::vec2 GetDeltaMousePos() = 0;
+			virtual glm::ivec2 GetGlobalMousePos() = 0;
+			virtual glm::ivec2 GetDeltaMousePos() = 0;
 
 		};
 	}
